Fixed mouse motion -> should be independent of screen
Former-commit-id: 7817cd99c33a9f4211b7e77022153b31d6634e0d

+++ include/events/ViewChangedEvent.hpp
@@ -6,8 +6,10 @@
 
 namespace sw {
     struct ViewChangedEvent {
-        ViewChangedEvent(int x, int y) : delta_up_(y), delta_right_(x) {}
+        ViewChangedEvent(float x, float y) : delta_up_(y), delta_right_(x) {}
 
+        // The value in these represents the ratio of 'pixels mouse moved' / 'window width'
+        // Scale accordingly! And safely...
         float delta_up_, delta_right_;
     };
 }
